refactor(desktop): the toolset is called desktop, not live_view
A "live view" was one agent-driven component in the chat sidebar. What
the toolset became is the machine's screen: which windows are open, what
each shows, the shared browser, the data server behind all of them. Named
after its first feature, every error message about the desktop mentioned
a thing the user had never opened — "怎么在用live view" is the whole
report, and it is a fair question.

So: pantheon/toolsets/live_view → toolsets/desktop, LiveViewToolSet →
DesktopToolSet, and the default toolset name follows.

Two aliases, because the pod and the browser ship separately and a rename
is live for one of them before the other:

  - endpoint routing maps "live_view" → "desktop" (TOOLSET_ALIASES), so a
    browser that has not been redeployed keeps working instead of meeting
    "Toolset 'live_view' not found in endpoint services" on every call —
    which is a desktop where nothing opens at all;
  - LiveViewToolSet stays bound to the class, so a saved agent config
    naming it still resolves.

The live_view_* tools themselves are untouched here; retiring them is its
own change, and it has to rewrite eleven app skills that still teach them.
