+++ tests/test_desktop_endpoint_management.py
@@ -4,7 +4,7 @@ import httpx
 import pytest
 from aiohttp import web
 
-from pantheon.toolsets.live_view.data_server import LiveViewDataServer
+from pantheon.toolsets.desktop.data_server import LiveViewDataServer
 
 
 @pytest.mark.asyncio
